Trying a different variation of the noarch builds
I think the issue comes from pinning Python in the Host and the Run section,
which appears to cause a conflict in building the package such that the
package itself cannot be found.

I'm basing this PR on my local testing and what I am seeing from the
rpcq-feedstock which appears to be the only other package which
uses the dataclasses package on conda-forge and builds on noarch.
Or, at least, they did.

Possible replacement to \#4, but experimental.